Split orchestration files to match contents
Move the Runs tracker into its own runs.go, and rename task.go to taskstop.go
(which now holds only the TaskStop tool). The shared substrate and the control
tool no longer share a file. agent.go and monitor.go are unchanged.
